Randomize the name used in the websocket registration
I and others have seen situations where subsequent websocket connections are not permitted and a WebSocketBadStatusExeption is raised if there are some non-cleanly terminated sessions. I suspect there must be an upper limit on the number of simultaneous webesockets with the same name. As we are using an undocumented feature this is a best guess. This patch randomizes the name on startup to avoid this issue.